func_alma_notifier/requirements.txt
-e file:///Users/trb74/azure-functions/wrlc/barcode_checks/shared_code ; python_version >= "3.11" and python_version < "4.0"
alembic==1.15.2 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
azure-common==1.1.28 ; python_version >= "3.11" and python_version < "4.0"
azure-communication-email==1.0.1b1 ; python_version >= "3.11" and python_version < "4.0"
azure-core==1.33.0 ; python_version >= "3.11" and python_version < "4.0"
azure-functions==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
azure-identity==1.21.0 ; python_version >= "3.11" and python_version < "4.0"
azure-mgmt-core==1.5.0 ; python_version >= "3.11" and python_version < "4.0"
azure-storage-blob==12.25.1 ; python_version >= "3.11" and python_version < "4.0"
azure-storage-queue==12.12.0 ; python_version >= "3.11" and python_version < "4.0"
backports-tarfile==1.2.0 ; python_version == "3.11"
build==1.2.2.post1 ; python_version >= "3.11" and python_version < "4.0"
cachecontrol==0.14.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and os_name == "nt"
crashtest==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
cryptography==44.0.2 ; python_version >= "3.11" and python_version < "4.0"
distlib==0.3.9 ; python_version >= "3.11" and python_version < "4.0"
dulwich==0.22.8 ; python_version >= "3.11" and python_version < "4.0"
fastjsonschema==2.21.1 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.18.0 ; python_version >= "3.11" and python_version < "4.0"
findpython==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.0 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.8 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version == "3.11"
installer==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
isodate==0.7.2 ; python_version >= "3.11" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
jaraco-context==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
jaraco-functools==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
jeepney==0.9.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
keyring==25.6.0 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
more-itertools==10.6.0 ; python_version >= "3.11" and python_version < "4.0"
msal-extensions==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
msal==1.32.0 ; python_version >= "3.11" and python_version < "4.0"
msgpack==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
msrest==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.2.4 ; python_version >= "3.11" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
pbs-installer==2025.4.9 ; python_version >= "3.11" and python_version < "4.0"
pkginfo==1.12.1.2 ; python_version >= "3.11" and python_version < "4.0"
platformdirs==4.3.7 ; python_version >= "3.11" and python_version < "4.0"
poetry-core==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
poetry-plugin-export==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
poetry==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pymysql==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
pyproject-hooks==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
pywin32-ctypes==0.2.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
rapidfuzz==3.13.0 ; python_version >= "3.11" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "linux"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.40 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.11" and python_version < "4.0"
trove-classifiers==2025.4.11.15 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
virtualenv==20.30.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
xattr==1.1.4 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "darwin"
zipp==3.21.0 ; python_version == "3.11"
zstandard==0.23.0 ; python_version >= "3.11" and python_version < "4.0"
